Code style #2
